Fix announcements render logic
collections.accnouncement returns objects like this:

{ inputPath: './test1.md',
  fileSlug: 'test1', // fileSlug was added in 0.5.3
  outputPath: './_site/test1/index.html',
  url: '/test1/',
  date: 2018-01-09T04:10:17.000Z,
  data: { title: 'Test Title', tags: ['tag1', 'tag2'], date: 'Last Modified' },
  templateContent: '<h1>This is my title</h1>\n\n<p>This is content…' }

when using nunjucks selectattr(show) method before, it didn't work because the
show attribute is in the data field of the announcement object.

Rather than figuring out how to access a nested field with nunjucks, I decided
to write a custom filter to eleventy which also allows me to remove the tags field
from an announcement markdown file. Of coures when I add another collection with a
show: field, I will need to add tagging back in, but for now it's simpler.